refactor: Statistics 系モデルをアプリケーション層に移動 (#597)
UserMatchStatistics と CircleMatchStatistics はドメインの不変条件やビジネスルールを
持たない集計結果の型であり、アプリケーション層のサービスでのみ使用されているため、
server/domain/models/match/ から server/application/user/ に移動した。
